+++ changelog.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [UNRELEASED]
+* [backend] - Check if bot is still a mod on registered subs.  If it's not, remove the sub
 * [backend] - Moved background scheduled tasks to a new scheduler system
 * [bugfix] - Fixed a timing issue causing the bot to make duplicate replies